sparql interface integration
I added a directory static/templates/working_sparql_framework_in_docs/. If all of its content is added to the docs directory *after* all the other html and css pages have been created in the docs directory, and if the index navbar is modified as suggested in the "navbar_2.html" file in templates, the website should work correctly, linking the two pages.
I removed the styles.css file from the working_sparql_framework directory and addedd the additional rules to the original template for css, so to avoid potential problems in over-writing the automatically generated css from templates, as the two files would have the same name. It should not interfere with the rest of the code anyway. If so, feel free to remove the added rules from the original css file, and I'll re-add a different css in the working_sparql_framework with a different name, so that when added to "docs" it won't influence the functioning of index.html.
the file index_static_2.html was only added as a legacy file, since it was the version of the index.html file that used to work as final exposed version, together with sparql.html - both in the docs directory (so, some parts of the code, such as the navbar were modified manually after the automatic generation of the file, and it might help in reconstructing the behavior of the website)